+++ src/padsynth.c
@@ -210,7 +210,7 @@ padsynth_render(y_sample_t *sample)
     }
 
     /* calculate the output table size */
-    i = lrintf((float)global.sample_rate * 2.5f);  /* at least 2.5 seconds long -FIX- this should be configurable */
+    i = lrintf((float)sample->sample_rate * 2.5f);  /* at least 2.5 seconds long -FIX- this should be configurable */
     N = WAVETABLE_POINTS * 2;
     while (N < i) {
         if (N * 5 / 4 >= i) { N = N * 5 / 4; break; }
@@ -309,7 +309,7 @@ padsynth_render(y_sample_t *sample)
      * so we find a new 'samplerate' that will result in fi * N being exactly 778:
      *    samplerate = f * N / fc = 44076.8
      */
-    fc0 = lrintf(f / (float)global.sample_rate * (float)N);
+    fc0 = lrintf(f / (float)sample->sample_rate * (float)N);
     sample->period = (float)N / (float)fc0;  /* frames per period */
     samplerate = f * sample->period;
     /* YDB_MESSAGE(YDB_SAMPLE, " padsynth_render: size = %d, f = %f, fc0 = %d, period = %f\n", N, f, fc0, sample->period); */
@@ -327,7 +327,7 @@ padsynth_render(y_sample_t *sample)
      * upper partial limit:
      *   ((global.sample_rate / 2) * samplerate / global.sample_rate) / samplerate * N / shift
      */
-    plimit_low = lrintf(20.0f / (float)global.sample_rate * (float)N);
+    plimit_low = lrintf(20.0f / (float)sample->sample_rate * (float)N);
     /* plimit_high = lrintf(20000.0f / (float)global.sample_rate * (float)N / 1.25992f); */
     plimit_high = lrintf((float)N / 2 / 1.25992f);
     /* YDB_MESSAGE(YDB_SAMPLE, " padsynth_render: nominal rate = %f, plimit low = %d, plimit high = %d\n", samplerate, plimit_low, plimit_high); */
@@ -616,7 +616,7 @@ padsynth_oscillator(unsigned long sample_count, y_sosc_t *sosc,
             level_b += level_b_delta;
 
             pos += w * period;
-            if (pos >= length) pos -= length;
+            if (pos >= length) pos = fmod(pos, length);
             /* sampleset oscillators do not export sync */
         }
 
@@ -663,8 +663,8 @@ padsynth_oscillator(unsigned long sample_count, y_sosc_t *sosc,
 
             pos0 += w * period0;
             pos1 += w * period1;
-            if (pos0 >= length0) pos0 -= length0;
-            if (pos1 >= length1) pos1 -= length1;
+            if (pos0 >= length0) pos0 = fmod(pos0, length0);
+            if (pos1 >= length1) pos1 = fmod(pos1, length1);
             /* sampleset oscillators do not export sync */
         }
 
@@ -689,7 +689,7 @@ padsynth_oscillator(unsigned long sample_count, y_sosc_t *sosc,
          * it an multiple of the period length to minimize phase cancellation
          * when summed to mono */
         posr = posl + rint(length / 2.0 / (double)period) * (double)period;
-        if (posr >= length) posr -= length;
+        if (posr >= length) posr = fmod(posr, length);
 
         for (sample = 0; sample < sample_count; sample++) {
 
@@ -710,9 +710,9 @@ padsynth_oscillator(unsigned long sample_count, y_sosc_t *sosc,
             level_b += level_b_delta;
 
             posl += w * period;
-            if (posl >= length) posl -= length;
+            if (posl >= length) posl = fmod(posl, length);
             posr += w * period;
-            if (posr >= length) posr -= length;
+            if (posr >= length) posr = fmod(posr, length);
             /* sampleset oscillators do not export sync */
         }
 
@@ -740,8 +740,8 @@ padsynth_oscillator(unsigned long sample_count, y_sosc_t *sosc,
         if (posl1 >= length1) posl1 = 0.0;
         posr0 = posl0 + rint(length0 / 2.0 / (double)period0) * (double)period0;
         posr1 = posl1 + rint(length1 / 2.0 / (double)period1) * (double)period1;
-        if (posr0 >= length0) posr0 -= length0;
-        if (posr1 >= length1) posr1 -= length1;
+        if (posr0 >= length0) posr0 = fmod(posr0, length0);
+        if (posr1 >= length1) posr1 = fmod(posr1, length1);
 
         for (sample = 0; sample < sample_count; sample++) {
 
@@ -773,10 +773,10 @@ padsynth_oscillator(unsigned long sample_count, y_sosc_t *sosc,
             posr0 += w * period0;
             posl1 += w * period1;
             posr1 += w * period1;
-            if (posl0 >= length0) posl0 -= length0;
-            if (posr0 >= length0) posr0 -= length0;
-            if (posl1 >= length1) posl1 -= length1;
-            if (posr1 >= length1) posr1 -= length1;
+            if (posl0 >= length0) posl0 = fmod(posl0, length0);
+            if (posr0 >= length0) posr0 = fmod(posr0, length0);
+            if (posl1 >= length1) posl1 = fmod(posl1, length1);
+            if (posr1 >= length1) posr1 = fmod(posr1, length1);
             /* sampleset oscillators do not export sync */
         }